requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.6
requests==2.31.0
gunicorn==20.1.0
PyMuPDF==1.24.13
python-docx==1.1.2
spacy==3.8.2
channels==4.2.0
daphne==4.1.2
channels-redis==4.2.0
django-cors-headers==4.6.0
